refactor(api-gateway): 调整entity所在包
