+++ api-gateway/src/main/java/org/njuse17advancedse/apigateway/domains/repo/entity/PaperRepo.java
@@ -1,7 +1,7 @@
 package org.njuse17advancedse.apigateway.domains.repo.entity;
 
 import java.util.ArrayList;
-import org.njuse17advancedse.apigateway.domains.entity.Paper;
+import org.njuse17advancedse.apigateway.infra.entity.Paper;
 import org.springframework.stereotype.Repository;
 
 @Repository
