+++ api-gateway/src/main/java/org/njuse17advancedse/apigateway/infra/entity/Affiliation.java
@@ -1,4 +1,4 @@
-package org.njuse17advancedse.apigateway.domains.entity;
+package org.njuse17advancedse.apigateway.infra.entity;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
