+++ api-gateway/src/main/java/org/njuse17advancedse/apigateway/infra/entity/Journal.java
@@ -1,4 +1,4 @@
-package org.njuse17advancedse.apigateway.domains.entity;
+package org.njuse17advancedse.apigateway.infra.entity;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -8,12 +8,10 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Researcher {
+public class Journal {
   private String id;
 
   private String name;
-
-  private Affiliation affiliation;
 
   private List<Paper> papers;
 }
