+++ api-gateway/src/main/java/org/njuse17advancedse/apigateway/infra/entity/Researcher.java
@@ -1,4 +1,4 @@
-package org.njuse17advancedse.apigateway.domains.entity;
+package org.njuse17advancedse.apigateway.infra.entity;
 
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -8,12 +8,12 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class Domain {
+public class Researcher {
   private String id;
 
   private String name;
 
-  private List<Researcher> researchers;
+  private Affiliation affiliation;
 
   private List<Paper> papers;
 }
