+++ api-gateway/src/main/java/org/njuse17advancedse/apigateway/interfaces/dto/researcher/IPaper.java
@@ -4,9 +4,9 @@ import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.njuse17advancedse.apigateway.domains.entity.Domain;
-import org.njuse17advancedse.apigateway.domains.entity.Paper;
-import org.njuse17advancedse.apigateway.domains.entity.Researcher;
+import org.njuse17advancedse.apigateway.infra.entity.Domain;
+import org.njuse17advancedse.apigateway.infra.entity.Paper;
+import org.njuse17advancedse.apigateway.infra.entity.Researcher;
 
 @Data
 @AllArgsConstructor
