+++ api-gateway/src/main/java/org/njuse17advancedse/apigateway/interfaces/dto/researcher/IResearcher.java
@@ -4,8 +4,8 @@ import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.njuse17advancedse.apigateway.domains.entity.Affiliation;
-import org.njuse17advancedse.apigateway.domains.entity.Paper;
+import org.njuse17advancedse.apigateway.infra.entity.Affiliation;
+import org.njuse17advancedse.apigateway.infra.entity.Paper;
 
 @Data
 @AllArgsConstructor
